// frontend/ticket-booking-web/src/app/api/model/order-status.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const OrderStatus = {
    Pending: 'Pending',
    Paid: 'Paid',
    Cancelled: 'Cancelled'
} as const;
export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];