// frontend/ticket-booking-web/src/app/api/model/paged-result-of-order-summary-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderSummaryDto } from './order-summary-dto';


export interface PagedResultOfOrderSummaryDto { 
    items: Array<OrderSummaryDto>;
    page: number;
    pageSize: number;
    totalCount: number;
    totalPages?: number;
}