// frontend/ticket-booking-web/src/app/api/model/sort-direction.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SortDirection = {
    Asc: 'Asc',
    Desc: 'Desc'
} as const;
export type SortDirection = typeof SortDirection[keyof typeof SortDirection];